Reconfigure MARC entry
MARC is the metadata standard, MARCXML is a serialization of it.

Disciplines given in incorrect format. New subject page required. This
is a tricky one as MARC is used almost exclusively by the Library
community, but is used to describe resources from any domain.

MODS and MADS are other (metadata) XML schemas, not vocabularies.

NDMSO is part of LoC, and should not be listed separately.

Contact should be name and email address, not postal address.